Moved file read/writes to TestApp.java
App.java is the entry point for the program, so we'll use TestApp.java to handle calls to it and ensuring inputs and outputs are recording to specific files for debugging purposes.
