+++ nvcs_java/src/main/java/nvcs/App.java
@@ -1,14 +1,5 @@
 package nvcs;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
 import nvcs_components.Classifier;
 import nvcs_components.Node;
 import nvcs_components.Plot;
@@ -19,47 +10,9 @@ import nvcs_utilities.JsonRow;
 
 public class App 
 {
-    public static void main(String[] args) {
-        String type = args[0];
-        String inputPath = args[1];
-        String outputPath = args[2];
-        List<String> outputLines = new ArrayList<>();
+
+    public static Integer[] nvcs_classify(String type, String json) {
         Classifier classifier = getClassifierByType(type);
-
-        try (Stream<String> stream = Files.lines(Paths.get(inputPath))) {
-            stream.forEach(inLine -> {
-                String[] split = inLine.split("\t");
-                String ident = split[0];
-                String json = split[1];
-                Integer[] path = classify(classifier, json);
-                Integer solution = path.length > 0 ? path[path.length - 1] : -1;
-                if (path[path.length - 1] == -1)
-                    path = Arrays.copyOf(path, path.length - 1);
-
-                String joinedPath = Stream.of(path).map(String::valueOf).collect(Collectors.joining(", "));
-                String outputLine = String.format("%s\t%s\t[%s]", ident, solution, joinedPath);
-                outputLines.add(outputLine);
-            });
-        }
-        catch (IOException ex) {
-            System.err.println("Error reading file: " + ex);
-        }
-
-        try {
-            Files.write(Paths.get(outputPath), outputLines);
-        }
-        catch (Exception ex) {
-            System.err.println("Error writing file: " + ex);
-        }
-    }
-
-    public static Integer[] classifyWest(String json) {
-        Classifier classifier = getClassifierByType("west");
-        return classify(classifier, json);
-    }
-
-    public static Integer[] classifyEast(String json) {
-        Classifier classifier = getClassifierByType("east");
         return classify(classifier, json);
     }
