+++ nvcs_java/src/main/java/nvcs_components/KeyValue.java
@@ -1,6 +1,7 @@
 package nvcs_components;
 
 public class KeyValue {
+    
     public String key;
     public String value;
